decoupled namespace configurators
