+++ libsql-server/src/namespace/configurator/replica.rs
@@ -2,29 +2,51 @@ use std::pin::Pin;
 use std::sync::Arc;
 
 use futures::Future;
+use hyper::Uri;
 use libsql_replication::rpc::replication::replication_log_client::ReplicationLogClient;
 use tokio::task::JoinSet;
+use tonic::transport::Channel;
 
+use crate::connection::config::DatabaseConfig;
+use crate::connection::connection_manager::InnerWalManager;
 use crate::connection::write_proxy::MakeWriteProxyConn;
 use crate::connection::MakeConnection;
 use crate::database::{Database, ReplicaDatabase};
 use crate::namespace::broadcasters::BroadcasterHandle;
+use crate::namespace::configurator::helpers::make_stats;
 use crate::namespace::meta_store::MetaStoreHandle;
-use crate::namespace::{Namespace, RestoreOption};
-use crate::namespace::{
-    make_stats, NamespaceConfig, NamespaceName, NamespaceStore, ResetCb, ResetOp,
-    ResolveNamespacePathFn,
-};
+use crate::namespace::{Namespace, NamespaceBottomlessDbIdInit, RestoreOption};
+use crate::namespace::{NamespaceName, NamespaceStore, ResetCb, ResetOp, ResolveNamespacePathFn};
 use crate::{DB_CREATE_TIMEOUT, DEFAULT_AUTO_CHECKPOINT};
 
-use super::ConfigureNamespace;
+use super::{BaseNamespaceConfig, ConfigureNamespace};
 
-pub struct ReplicaConfigurator;
+pub struct ReplicaConfigurator {
+    base: BaseNamespaceConfig,
+    channel: Channel,
+    uri: Uri,
+    make_wal_manager: Arc<dyn Fn() -> InnerWalManager + Sync + Send + 'static>,
+}
+
+impl ReplicaConfigurator {
+    pub fn new(
+        base: BaseNamespaceConfig,
+        channel: Channel,
+        uri: Uri,
+        make_wal_manager: Arc<dyn Fn() -> InnerWalManager + Sync + Send + 'static>,
+    ) -> Self {
+        Self {
+            base,
+            channel,
+            uri,
+            make_wal_manager,
+        }
+    }
+}
 
 impl ConfigureNamespace for ReplicaConfigurator {
     fn setup<'a>(
         &'a self,
-        config: &'a NamespaceConfig,
         meta_store_handle: MetaStoreHandle,
         restore_option: RestoreOption,
         name: &'a NamespaceName,
@@ -32,13 +54,12 @@ impl ConfigureNamespace for ReplicaConfigurator {
         resolve_attach_path: ResolveNamespacePathFn,
         store: NamespaceStore,
         broadcaster: BroadcasterHandle,
-    ) -> Pin<Box<dyn Future<Output = crate::Result<Namespace>> + Send + 'a>>
-    {
+    ) -> Pin<Box<dyn Future<Output = crate::Result<Namespace>> + Send + 'a>> {
         Box::pin(async move {
             tracing::debug!("creating replica namespace");
-            let db_path = config.base_path.join("dbs").join(name.as_str());
-            let channel = config.channel.clone().expect("bad replica config");
-            let uri = config.uri.clone().expect("bad replica config");
+            let db_path = self.base.base_path.join("dbs").join(name.as_str());
+            let channel = self.channel.clone();
+            let uri = self.uri.clone();
 
             let rpc_client = ReplicationLogClient::with_origin(channel.clone(), uri.clone());
             let client = crate::replication::replicator_client::Client::new(
@@ -48,45 +69,46 @@ impl ConfigureNamespace for ReplicaConfigurator {
                 meta_store_handle.clone(),
                 store.clone(),
             )
-                .await?;
+            .await?;
             let applied_frame_no_receiver = client.current_frame_no_notifier.subscribe();
             let mut replicator = libsql_replication::replicator::Replicator::new(
                 client,
                 db_path.join("data"),
                 DEFAULT_AUTO_CHECKPOINT,
-                config.encryption_config.clone(),
+                None,
             )
-                .await?;
+            .await?;
 
             tracing::debug!("try perform handshake");
             // force a handshake now, to retrieve the primary's current replication index
             match replicator.try_perform_handshake().await {
                 Err(libsql_replication::replicator::Error::Meta(
-                        libsql_replication::meta::Error::LogIncompatible,
+                    libsql_replication::meta::Error::LogIncompatible,
                 )) => {
                     tracing::error!(
                         "trying to replicate incompatible logs, reseting replica and nuking db dir"
                     );
                     std::fs::remove_dir_all(&db_path).unwrap();
-                    return self.setup(
-                        config,
-                        meta_store_handle,
-                        restore_option,
-                        name,
-                        reset,
-                        resolve_attach_path,
-                        store,
-                        broadcaster,
-                    )
+                    return self
+                        .setup(
+                            meta_store_handle,
+                            restore_option,
+                            name,
+                            reset,
+                            resolve_attach_path,
+                            store,
+                            broadcaster,
+                        )
                         .await;
-                    }
+                }
                 Err(e) => Err(e)?,
                 Ok(_) => (),
             }
 
             tracing::debug!("done performing handshake");
 
-            let primary_current_replicatio_index = replicator.client_mut().primary_replication_index;
+            let primary_current_replicatio_index =
+                replicator.client_mut().primary_replication_index;
 
             let mut join_set = JoinSet::new();
             let namespace = name.clone();
@@ -144,36 +166,35 @@ impl ConfigureNamespace for ReplicaConfigurator {
                 &db_path,
                 &mut join_set,
                 meta_store_handle.clone(),
-                config.stats_sender.clone(),
+                self.base.stats_sender.clone(),
                 name.clone(),
                 applied_frame_no_receiver.clone(),
-                config.encryption_config.clone(),
             )
-                .await?;
+            .await?;
 
             let connection_maker = MakeWriteProxyConn::new(
                 db_path.clone(),
-                config.extensions.clone(),
+                self.base.extensions.clone(),
                 channel.clone(),
                 uri.clone(),
                 stats.clone(),
                 broadcaster,
                 meta_store_handle.clone(),
                 applied_frame_no_receiver,
-                config.max_response_size,
-                config.max_total_response_size,
+                self.base.max_response_size,
+                self.base.max_total_response_size,
                 primary_current_replicatio_index,
-                config.encryption_config.clone(),
+                None,
                 resolve_attach_path,
-                config.make_wal_manager.clone(),
+                self.make_wal_manager.clone(),
             )
-                .await?
-                .throttled(
-                    config.max_concurrent_connections.clone(),
-                    Some(DB_CREATE_TIMEOUT),
-                    config.max_total_response_size,
-                    config.max_concurrent_requests,
-                );
+            .await?
+            .throttled(
+                self.base.max_concurrent_connections.clone(),
+                Some(DB_CREATE_TIMEOUT),
+                self.base.max_total_response_size,
+                self.base.max_concurrent_requests,
+            );
 
             Ok(Namespace {
                 tasks: join_set,
@@ -186,5 +207,36 @@ impl ConfigureNamespace for ReplicaConfigurator {
                 path: db_path.into(),
             })
         })
+    }
+
+    fn cleanup<'a>(
+        &'a self,
+        namespace: &'a NamespaceName,
+        _db_config: &DatabaseConfig,
+        _prune_all: bool,
+        _bottomless_db_id_init: NamespaceBottomlessDbIdInit,
+    ) -> Pin<Box<dyn Future<Output = crate::Result<()>> + Send + 'a>> {
+        Box::pin(async move {
+            let ns_path = self.base.base_path.join("dbs").join(namespace.as_str());
+            if ns_path.try_exists()? {
+                tracing::debug!("removing database directory: {}", ns_path.display());
+                tokio::fs::remove_dir_all(ns_path).await?;
+            }
+            Ok(())
+        })
+    }
+
+    fn fork<'a>(
+        &'a self,
+        _from_ns: &'a Namespace,
+        _from_config: MetaStoreHandle,
+        _to_ns: NamespaceName,
+        _to_config: MetaStoreHandle,
+        _timestamp: Option<chrono::prelude::NaiveDateTime>,
+        _store: NamespaceStore,
+    ) -> Pin<Box<dyn Future<Output = crate::Result<Namespace>> + Send + 'a>> {
+        Box::pin(std::future::ready(Err(crate::Error::Fork(
+            super::fork::ForkError::ForkReplica,
+        ))))
     }
 }
