+++ libsql-server/src/namespace/mod.rs
@@ -1,52 +1,24 @@
-use std::path::{Path, PathBuf};
-use std::sync::atomic::AtomicBool;
-use std::sync::{Arc, Weak};
+use std::path::Path;
+use std::sync::Arc;
 
-use anyhow::{Context as _, Error};
-use bottomless::replicator::Options;
-use broadcasters::BroadcasterHandle;
+use anyhow::Context as _;
 use bytes::Bytes;
 use chrono::NaiveDateTime;
-use enclose::enclose;
 use futures_core::{Future, Stream};
-use hyper::Uri;
-use libsql_sys::wal::Sqlite3WalManager;
-use libsql_sys::EncryptionConfig;
-use tokio::io::AsyncBufReadExt;
-use tokio::sync::{watch, Semaphore};
 use tokio::task::JoinSet;
-use tokio::time::Duration;
-use tokio_util::io::StreamReader;
-use tonic::transport::Channel;
 use uuid::Uuid;
 
 use crate::auth::parse_jwt_keys;
 use crate::connection::config::DatabaseConfig;
-use crate::connection::connection_manager::InnerWalManager;
-use crate::connection::libsql::{open_conn, MakeLibSqlConn};
-use crate::connection::{Connection as _, MakeConnection};
-use crate::database::{
-    Database, DatabaseKind, PrimaryConnection, PrimaryConnectionMaker,
-};
-use crate::error::LoadDumpError;
-use crate::replication::script_backup_manager::ScriptBackupManager;
-use crate::replication::{FrameNo, ReplicationLogger};
-use crate::schema::SchedulerHandle;
+use crate::connection::Connection as _;
+use crate::database::Database;
 use crate::stats::Stats;
-use crate::{
-    StatsSender, BLOCKING_RT, DB_CREATE_TIMEOUT, DEFAULT_AUTO_CHECKPOINT,
-};
 
-pub use fork::ForkError;
-
-use self::fork::{ForkTask, PointInTimeRestore};
 use self::meta_store::MetaStoreHandle;
 pub use self::name::NamespaceName;
-use self::replication_wal::{make_replication_wal_wrapper, ReplicationWalWrapper};
 pub use self::store::NamespaceStore;
 
 pub mod broadcasters;
-mod fork;
 pub mod meta_store;
 mod name;
 pub mod replication_wal;
@@ -101,51 +73,6 @@ impl Namespace {
         &self.name
     }
 
-    /// completely remove resources associated with the namespace
-    pub(crate) async fn cleanup(
-        ns_config: &NamespaceConfig,
-        name: &NamespaceName,
-        db_config: &DatabaseConfig,
-        prune_all: bool,
-        bottomless_db_id_init: NamespaceBottomlessDbIdInit,
-    ) -> crate::Result<()> {
-        let ns_path = ns_config.base_path.join("dbs").join(name.as_str());
-        match ns_config.db_kind {
-            DatabaseKind::Primary => {
-                if let Some(ref options) = ns_config.bottomless_replication {
-                    let bottomless_db_id = match bottomless_db_id_init {
-                        NamespaceBottomlessDbIdInit::Provided(db_id) => db_id,
-                        NamespaceBottomlessDbIdInit::FetchFromConfig => {
-                            NamespaceBottomlessDbId::from_config(&db_config)
-                        }
-                    };
-                    let options = make_bottomless_options(options, bottomless_db_id, name.clone());
-                    let replicator = bottomless::replicator::Replicator::with_options(
-                        ns_path.join("data").to_str().unwrap(),
-                        options,
-                    )
-                    .await?;
-                    if prune_all {
-                        let delete_all = replicator.delete_all(None).await?;
-                        // perform hard deletion in the background
-                        tokio::spawn(delete_all.commit());
-                    } else {
-                        // for soft delete make sure that local db is fully backed up
-                        replicator.savepoint().confirmed().await?;
-                    }
-                }
-            }
-            DatabaseKind::Replica => (),
-        }
-
-        if ns_path.try_exists()? {
-            tracing::debug!("removing database directory: {}", ns_path.display());
-            tokio::fs::remove_dir_all(ns_path).await?;
-        }
-
-        Ok(())
-    }
-
     async fn destroy(mut self) -> anyhow::Result<()> {
         self.tasks.shutdown().await;
         self.db.destroy();
@@ -195,292 +122,10 @@ impl Namespace {
     pub fn config_changed(&self) -> impl Future<Output = ()> {
         self.db_config_store.changed()
     }
-
-    #[tracing::instrument(skip_all)]
-    async fn make_primary_connection_maker(
-        ns_config: &NamespaceConfig,
-        meta_store_handle: &MetaStoreHandle,
-        db_path: &Path,
-        name: &NamespaceName,
-        restore_option: RestoreOption,
-        block_writes: Arc<AtomicBool>,
-        join_set: &mut JoinSet<anyhow::Result<()>>,
-        resolve_attach_path: ResolveNamespacePathFn,
-        broadcaster: BroadcasterHandle,
-    ) -> crate::Result<(PrimaryConnectionMaker, ReplicationWalWrapper, Arc<Stats>)> {
-        let db_config = meta_store_handle.get();
-        let bottomless_db_id = NamespaceBottomlessDbId::from_config(&db_config);
-        // FIXME: figure how to to it per-db
-        let mut is_dirty = {
-            let sentinel_path = db_path.join(".sentinel");
-            if sentinel_path.try_exists()? {
-                true
-            } else {
-                tokio::fs::File::create(&sentinel_path).await?;
-                false
-            }
-        };
-
-        // FIXME: due to a bug in logger::checkpoint_db we call regular checkpointing code
-        // instead of our virtual WAL one. It's a bit tangled to fix right now, because
-        // we need WAL context for checkpointing, and WAL context needs the ReplicationLogger...
-        // So instead we checkpoint early, *before* bottomless gets initialized. That way
-        // we're sure bottomless won't try to back up any existing WAL frames and will instead
-        // treat the existing db file as the source of truth.
-
-        let bottomless_replicator = match ns_config.bottomless_replication {
-            Some(ref options) => {
-                tracing::debug!("Checkpointing before initializing bottomless");
-                crate::replication::primary::logger::checkpoint_db(&db_path.join("data"))?;
-                tracing::debug!("Checkpointed before initializing bottomless");
-                let options = make_bottomless_options(options, bottomless_db_id, name.clone());
-                let (replicator, did_recover) =
-                    init_bottomless_replicator(db_path.join("data"), options, &restore_option)
-                        .await?;
-                tracing::debug!("Completed init of bottomless replicator");
-                is_dirty |= did_recover;
-                Some(replicator)
-            }
-            None => None,
-        };
-
-        tracing::debug!("Checking fresh db");
-        let is_fresh_db = check_fresh_db(&db_path)?;
-        // switch frame-count checkpoint to time-based one
-        let auto_checkpoint = if ns_config.checkpoint_interval.is_some() {
-            0
-        } else {
-            DEFAULT_AUTO_CHECKPOINT
-        };
-
-        let logger = Arc::new(ReplicationLogger::open(
-            &db_path,
-            ns_config.max_log_size,
-            ns_config.max_log_duration,
-            is_dirty,
-            auto_checkpoint,
-            ns_config.scripted_backup.clone(),
-            name.clone(),
-            ns_config.encryption_config.clone(),
-        )?);
-
-        tracing::debug!("sending stats");
-
-        let stats = make_stats(
-            &db_path,
-            join_set,
-            meta_store_handle.clone(),
-            ns_config.stats_sender.clone(),
-            name.clone(),
-            logger.new_frame_notifier.subscribe(),
-            ns_config.encryption_config.clone(),
-        )
-        .await?;
-
-        tracing::debug!("Making replication wal wrapper");
-        let wal_wrapper = make_replication_wal_wrapper(bottomless_replicator, logger.clone());
-
-        tracing::debug!("Opening libsql connection");
-
-        let connection_maker = MakeLibSqlConn::new(
-            db_path.to_path_buf(),
-            wal_wrapper.clone(),
-            stats.clone(),
-            broadcaster,
-            meta_store_handle.clone(),
-            ns_config.extensions.clone(),
-            ns_config.max_response_size,
-            ns_config.max_total_response_size,
-            auto_checkpoint,
-            logger.new_frame_notifier.subscribe(),
-            ns_config.encryption_config.clone(),
-            block_writes,
-            resolve_attach_path,
-            ns_config.make_wal_manager.clone(),
-        )
-        .await?
-        .throttled(
-            ns_config.max_concurrent_connections.clone(),
-            Some(DB_CREATE_TIMEOUT),
-            ns_config.max_total_response_size,
-            ns_config.max_concurrent_requests,
-        );
-
-        tracing::debug!("Completed opening libsql connection");
-
-        // this must happen after we create the connection maker. The connection maker old on a
-        // connection to ensure that no other connection is closing while we try to open the dump.
-        // that would cause a SQLITE_LOCKED error.
-        match restore_option {
-            RestoreOption::Dump(_) if !is_fresh_db => {
-                Err(LoadDumpError::LoadDumpExistingDb)?;
-            }
-            RestoreOption::Dump(dump) => {
-                let conn = connection_maker.create().await?;
-                tracing::debug!("Loading dump");
-                load_dump(dump, conn).await?;
-                tracing::debug!("Done loading dump");
-            }
-            _ => { /* other cases were already handled when creating bottomless */ }
-        }
-
-        join_set.spawn(run_periodic_compactions(logger.clone()));
-
-        tracing::debug!("Done making primary connection");
-
-        Ok((connection_maker, wal_wrapper, stats))
-    }
-
-    async fn fork(
-        ns_config: &NamespaceConfig,
-        from_ns: &Namespace,
-        from_config: MetaStoreHandle,
-        to_ns: NamespaceName,
-        to_config: MetaStoreHandle,
-        timestamp: Option<NaiveDateTime>,
-        store: NamespaceStore,
-    ) -> crate::Result<Namespace> {
-        let from_config = from_config.get();
-        match ns_config.db_kind {
-            DatabaseKind::Primary => {
-                let bottomless_db_id = NamespaceBottomlessDbId::from_config(&from_config);
-                let restore_to = if let Some(timestamp) = timestamp {
-                    if let Some(ref options) = ns_config.bottomless_replication {
-                        Some(PointInTimeRestore {
-                            timestamp,
-                            replicator_options: make_bottomless_options(
-                                options,
-                                bottomless_db_id.clone(),
-                                from_ns.name().clone(),
-                            ),
-                        })
-                    } else {
-                        return Err(crate::Error::Fork(ForkError::BackupServiceNotConfigured));
-                    }
-                } else {
-                    None
-                };
-
-                let logger = match &from_ns.db {
-                    Database::Primary(db) => db.wal_wrapper.wrapper().logger(),
-                    Database::Schema(db) => db.wal_wrapper.wrapper().logger(),
-                    _ => {
-                        return Err(crate::Error::Fork(ForkError::Internal(Error::msg(
-                            "Invalid source database type for fork",
-                        ))));
-                    }
-                };
-
-                let fork_task = ForkTask {
-                    base_path: ns_config.base_path.clone(),
-                    to_namespace: to_ns.clone(),
-                    logger,
-                    restore_to,
-                    to_config,
-                    store,
-                };
-
-                let ns = fork_task.fork().await?;
-                Ok(ns)
-            }
-            DatabaseKind::Replica => Err(ForkError::ForkReplica.into()),
-        }
-    }
-}
-
-pub struct NamespaceConfig {
-    /// Default database kind the store should be Creating
-    pub(crate) db_kind: DatabaseKind,
-    // Common config
-    pub(crate) base_path: Arc<Path>,
-    pub(crate) max_log_size: u64,
-    pub(crate) max_log_duration: Option<Duration>,
-    pub(crate) extensions: Arc<[PathBuf]>,
-    pub(crate) stats_sender: StatsSender,
-    pub(crate) max_response_size: u64,
-    pub(crate) max_total_response_size: u64,
-    pub(crate) checkpoint_interval: Option<Duration>,
-    pub(crate) max_concurrent_connections: Arc<Semaphore>,
-    pub(crate) max_concurrent_requests: u64,
-    pub(crate) encryption_config: Option<EncryptionConfig>,
-
-    // Replica specific config
-    /// grpc channel for replica
-    pub channel: Option<Channel>,
-    /// grpc uri
-    pub uri: Option<Uri>,
-
-    // primary only config
-    pub(crate) bottomless_replication: Option<bottomless::replicator::Options>,
-    pub(crate) scripted_backup: Option<ScriptBackupManager>,
-    pub(crate) migration_scheduler: SchedulerHandle,
-    pub(crate) make_wal_manager: Arc<dyn Fn() -> InnerWalManager + Sync + Send + 'static>,
 }
 
 pub type DumpStream =
     Box<dyn Stream<Item = std::io::Result<Bytes>> + Send + Sync + 'static + Unpin>;
-
-fn make_bottomless_options(
-    options: &Options,
-    namespace_db_id: NamespaceBottomlessDbId,
-    name: NamespaceName,
-) -> Options {
-    let mut options = options.clone();
-    let mut db_id = match namespace_db_id {
-        NamespaceBottomlessDbId::Namespace(id) => id,
-        // FIXME(marin): I don't like that, if bottomless is enabled, proper config must be passed.
-        NamespaceBottomlessDbId::NotProvided => options.db_id.unwrap_or_default(),
-    };
-
-    db_id = format!("ns-{db_id}:{name}");
-    options.db_id = Some(db_id);
-    options
-}
-
-async fn make_stats(
-    db_path: &Path,
-    join_set: &mut JoinSet<anyhow::Result<()>>,
-    meta_store_handle: MetaStoreHandle,
-    stats_sender: StatsSender,
-    name: NamespaceName,
-    mut current_frame_no: watch::Receiver<Option<FrameNo>>,
-    encryption_config: Option<EncryptionConfig>,
-) -> anyhow::Result<Arc<Stats>> {
-    tracing::debug!("creating stats type");
-    let stats = Stats::new(name.clone(), db_path, join_set).await?;
-
-    // the storage monitor is optional, so we ignore the error here.
-    tracing::debug!("stats created, sending stats");
-    let _ = stats_sender
-        .send((name.clone(), meta_store_handle, Arc::downgrade(&stats)))
-        .await;
-
-    join_set.spawn({
-        let stats = stats.clone();
-        // initialize the current_frame_no value
-        current_frame_no
-            .borrow_and_update()
-            .map(|fno| stats.set_current_frame_no(fno));
-        async move {
-            while current_frame_no.changed().await.is_ok() {
-                current_frame_no
-                    .borrow_and_update()
-                    .map(|fno| stats.set_current_frame_no(fno));
-            }
-            Ok(())
-        }
-    });
-
-    join_set.spawn(run_storage_monitor(
-        db_path.into(),
-        Arc::downgrade(&stats),
-        encryption_config,
-    ));
-
-    tracing::debug!("done sending stats, and creating bg tasks");
-
-    Ok(stats)
-}
 
 #[derive(Default)]
 pub enum RestoreOption {
@@ -494,190 +139,4 @@ pub enum RestoreOption {
     /// Restore database state to a backup version present at a specific point in time.
     /// Granularity depends of how frequently WAL log pages are being snapshotted.
     PointInTime(NaiveDateTime),
-}
-
-const WASM_TABLE_CREATE: &str =
-    "CREATE TABLE libsql_wasm_func_table (name text PRIMARY KEY, body text) WITHOUT ROWID;";
-
-async fn load_dump<S>(dump: S, conn: PrimaryConnection) -> crate::Result<(), LoadDumpError>
-where
-    S: Stream<Item = std::io::Result<Bytes>> + Unpin,
-{
-    let mut reader = tokio::io::BufReader::new(StreamReader::new(dump));
-    let mut curr = String::new();
-    let mut line = String::new();
-    let mut skipped_wasm_table = false;
-    let mut n_stmt = 0;
-    let mut line_id = 0;
-
-    while let Ok(n) = reader.read_line(&mut curr).await {
-        line_id += 1;
-        if n == 0 {
-            break;
-        }
-        let trimmed = curr.trim();
-        if trimmed.is_empty() || trimmed.starts_with("--") {
-            curr.clear();
-            continue;
-        }
-        // FIXME: it's well known bug that comment ending with semicolon will be handled incorrectly by currend dump processing code
-        let statement_end = trimmed.ends_with(';');
-
-        // we want to concat original(non-trimmed) lines as trimming will join all them in one
-        // single-line statement which is incorrect if comments in the end are present
-        line.push_str(&curr);
-        curr.clear();
-
-        // This is a hack to ignore the libsql_wasm_func_table table because it is already created
-        // by the system.
-        if !skipped_wasm_table && line.trim() == WASM_TABLE_CREATE {
-            skipped_wasm_table = true;
-            line.clear();
-            continue;
-        }
-
-        if statement_end {
-            n_stmt += 1;
-            // dump must be performd within a txn
-            if n_stmt > 2 && conn.is_autocommit().await.unwrap() {
-                return Err(LoadDumpError::NoTxn);
-            }
-
-            line = tokio::task::spawn_blocking({
-                let conn = conn.clone();
-                move || -> crate::Result<String, LoadDumpError> {
-                    conn.with_raw(|conn| conn.execute(&line, ())).map_err(|e| {
-                        LoadDumpError::Internal(format!("line: {}, error: {}", line_id, e))
-                    })?;
-                    Ok(line)
-                }
-            })
-            .await??;
-            line.clear();
-        } else {
-            line.push(' ');
-        }
-    }
-    tracing::debug!("loaded {} lines from dump", line_id);
-
-    if !conn.is_autocommit().await.unwrap() {
-        tokio::task::spawn_blocking({
-            let conn = conn.clone();
-            move || -> crate::Result<(), LoadDumpError> {
-                conn.with_raw(|conn| conn.execute("rollback", ()))?;
-                Ok(())
-            }
-        })
-        .await??;
-        return Err(LoadDumpError::NoCommit);
-    }
-
-    Ok(())
-}
-
-pub async fn init_bottomless_replicator(
-    path: impl AsRef<std::path::Path>,
-    options: bottomless::replicator::Options,
-    restore_option: &RestoreOption,
-) -> anyhow::Result<(bottomless::replicator::Replicator, bool)> {
-    tracing::debug!("Initializing bottomless replication");
-    let path = path
-        .as_ref()
-        .to_str()
-        .ok_or_else(|| anyhow::anyhow!("Invalid db path"))?
-        .to_owned();
-    let mut replicator = bottomless::replicator::Replicator::with_options(path, options).await?;
-
-    let (generation, timestamp) = match restore_option {
-        RestoreOption::Latest | RestoreOption::Dump(_) => (None, None),
-        RestoreOption::Generation(generation) => (Some(*generation), None),
-        RestoreOption::PointInTime(timestamp) => (None, Some(*timestamp)),
-    };
-
-    let (action, did_recover) = replicator.restore(generation, timestamp).await?;
-    match action {
-        bottomless::replicator::RestoreAction::SnapshotMainDbFile => {
-            replicator.new_generation().await;
-            if let Some(_handle) = replicator.snapshot_main_db_file(true).await? {
-                tracing::trace!("got snapshot handle after restore with generation upgrade");
-            }
-            // Restoration process only leaves the local WAL file if it was
-            // detected to be newer than its remote counterpart.
-            replicator.maybe_replicate_wal().await?
-        }
-        bottomless::replicator::RestoreAction::ReuseGeneration(gen) => {
-            replicator.set_generation(gen);
-        }
-    }
-
-    Ok((replicator, did_recover))
-}
-
-async fn run_periodic_compactions(logger: Arc<ReplicationLogger>) -> anyhow::Result<()> {
-    // calling `ReplicationLogger::maybe_compact()` is cheap if the compaction does not actually
-    // take place, so we can afford to poll it very often for simplicity
-    let mut interval = tokio::time::interval(tokio::time::Duration::from_millis(1000));
-    interval.set_missed_tick_behavior(tokio::time::MissedTickBehavior::Delay);
-
-    loop {
-        interval.tick().await;
-        let handle = BLOCKING_RT.spawn_blocking(enclose! {(logger) move || {
-            logger.maybe_compact()
-        }});
-        handle
-            .await
-            .expect("Compaction task crashed")
-            .context("Compaction failed")?;
-    }
-}
-
-fn check_fresh_db(path: &Path) -> crate::Result<bool> {
-    let is_fresh = !path.join("wallog").try_exists()?;
-    Ok(is_fresh)
-}
-
-// Periodically check the storage used by the database and save it in the Stats structure.
-// TODO: Once we have a separate fiber that does WAL checkpoints, running this routine
-// right after checkpointing is exactly where it should be done.
-async fn run_storage_monitor(
-    db_path: PathBuf,
-    stats: Weak<Stats>,
-    encryption_config: Option<EncryptionConfig>,
-) -> anyhow::Result<()> {
-    // on initialization, the database file doesn't exist yet, so we wait a bit for it to be
-    // created
-    tokio::time::sleep(Duration::from_secs(1)).await;
-
-    let duration = tokio::time::Duration::from_secs(60);
-    let db_path: Arc<Path> = db_path.into();
-    loop {
-        let db_path = db_path.clone();
-        let Some(stats) = stats.upgrade() else {
-            return Ok(());
-        };
-
-        let encryption_config = encryption_config.clone();
-        let _ = tokio::task::spawn_blocking(move || {
-            // because closing the last connection interferes with opening a new one, we lazily
-            // initialize a connection here, and keep it alive for the entirety of the program. If we
-            // fail to open it, we wait for `duration` and try again later.
-            match open_conn(&db_path, Sqlite3WalManager::new(), Some(rusqlite::OpenFlags::SQLITE_OPEN_READ_ONLY), encryption_config) {
-                Ok(mut conn) => {
-                    if let Ok(tx) = conn.transaction() {
-                        let page_count = tx.query_row("pragma page_count;", [], |row| { row.get::<usize, u64>(0) });
-                        let freelist_count = tx.query_row("pragma freelist_count;", [], |row| { row.get::<usize, u64>(0) });
-                        if let (Ok(page_count), Ok(freelist_count)) = (page_count, freelist_count) {
-                            let storage_bytes_used = (page_count - freelist_count) * 4096;
-                            stats.set_storage_bytes_used(storage_bytes_used);
-                        }
-                    }
-                },
-                Err(e) => {
-                    tracing::warn!("failed to open connection for storager monitor: {e}, trying again in {duration:?}");
-                },
-            }
-        }).await;
-
-        tokio::time::sleep(duration).await;
-    }
 }
